+++ CHANGELOG.md
@@ -34,6 +34,8 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Changed
 
+- Failed DNS responses are now shown as errors in the query log ([#4136]).
+
 - The `edge` channel has been switched to the new UI and versioning scheme.
 
 ### Deprecated
@@ -44,6 +46,7 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 - Blocked requests without an EDNS(0) OPT record ([#8183]).
 
+[#4136]: https://github.com/AdguardTeam/AdGuardHome/issues/4136
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
 
 <!--